# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DICOM Gateway API
  description: |
    REST API for DICOM Gateway - A lightweight, HIPAA-compliant Linux DICOM Gateway
    for receiving and forwarding DICOM studies with byte-preserving integrity.
    
    ## Features
    - C-STORE SCP (receiver) and SCU (forwarder) with byte-preserving transmission
    - PostgreSQL-backed job queue with LISTEN/NOTIFY
    - Role-based access control (RBAC)
    - Comprehensive audit logging
    - Prometheus metrics export
    - Configuration management via YAML
  version: 0.1.0
  contact:
    name: DICOM Gateway Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://gateway.example.com/api/v1
    description: Production server
...
paths:
...
  /destinations:
    get:
      summary: List destinations
      description: List all DICOM forwarding destinations
      tags:
        - Destinations
      security:
        - BearerAuth: []
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
          description: Filter by enabled status
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DestinationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create destination
      description: Create a new DICOM forwarding destination
      tags:
        - Destinations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DestinationCreate'
      responses:
        '201':
          description: Destination created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /destinations/{destination_id}:
    get:
      summary: Get destination
      description: Get destination by ID
      tags:
        - Destinations
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DestinationId'
      responses:
        '200':
          description: Destination details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update destination
      description: Update destination configuration
      tags:
        - Destinations
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DestinationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DestinationUpdate'
      responses:
        '200':
          description: Destination updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete destination
      description: Delete destination
      tags:
        - Destinations
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DestinationId'
      responses:
        '204':
          description: Destination deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /queues/stats:
    get:
      summary: Get queue statistics
      description: Get job queue statistics
      tags:
        - Queues
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Queue statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queues/retry:
    post:
      summary: Retry jobs
      description: Retry failed or dead-letter jobs
      tags:
        - Queues
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryRequest'
      responses:
        '200':
          description: Jobs retried
          content:
            application/json:
              schema:
                type: object
                properties:
                  retried:
                    type: integer
                  job_ids:
                    type: array
                    items:
                      type: string
                      format: uuid
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No jobs found to retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /queues/replay/{study_instance_uid}:
    post:
      summary: Replay study
      description: Create new forward jobs for a study
      tags:
        - Queues
      security:
        - BearerAuth: []
      parameters:
        - name: study_instance_uid
          in: path
          required: true
          schema:
            type: string
          description: Study Instance UID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                destination_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Study replayed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  jobs_created:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Skip:
      name: skip
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of records to skip
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      description: Maximum number of records to return
...
    DestinationId:
      name: destination_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Destination ID
  schemas:
...
    DestinationCreate:
      type: object
      required:
        - name
        - ae_title
        - host
        - port
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        ae_title:
          type: string
          minLength: 1
          maxLength: 255
        host:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        max_pdu:
          type: integer
          minimum: 128
          default: 16384
        timeout:
          type: integer
          minimum: 1
          default: 30
        connection_timeout:
          type: integer
          minimum: 1
          default: 10
        tls_enabled:
          type: boolean
          default: false
        tls_cert_path:
          type: string
          nullable: true
        tls_key_path:
          type: string
          nullable: true
        tls_ca_path:
          type: string
          nullable: true
        tls_no_verify:
          type: boolean
          default: false
        forwarding_rules:
          type: object
          nullable: true
        description:
          type: string
          nullable: true
    DestinationUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        ae_title:
          type: string
          minLength: 1
          maxLength: 255
        host:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        max_pdu:
          type: integer
          minimum: 128
        timeout:
          type: integer
          minimum: 1
        connection_timeout:
          type: integer
          minimum: 1
        tls_enabled:
          type: boolean
        tls_cert_path:
          type: string
          nullable: true
        tls_key_path:
          type: string
          nullable: true
        tls_ca_path:
          type: string
          nullable: true
        tls_no_verify:
          type: boolean
        forwarding_rules:
          type: object
          nullable: true
        description:
          type: string
          nullable: true
        enabled:
          type: boolean
    DestinationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ae_title:
          type: string
        host:
          type: string
        port:
          type: integer
        max_pdu:
          type: integer
        timeout:
          type: integer
        connection_timeout:
          type: integer
        tls_enabled:
          type: boolean
        tls_cert_path:
          type: string
          nullable: true
        tls_key_path:
          type: string
          nullable: true
        tls_ca_path:
          type: string
          nullable: true
        tls_no_verify:
          type: boolean
        enabled:
          type: boolean
        last_success_at:
          type: string
          format: date-time
          nullable: true
        last_failure_at:
          type: string
          format: date-time
          nullable: true
        consecutive_failures:
          type: integer
        forwarding_rules:
          type: object
          nullable: true
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    QueueStatsResponse:
      type: object
      properties:
        pending:
          type: integer
        processing:
          type: integer
        completed:
          type: integer
        failed:
          type: integer
        dead_letter:
          type: integer
    RetryRequest:
      type: object
      properties:
        job_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'